requirements.txt
beautifulsoup4==4.12.3
black==24.4.0
gradio==4.27.0
isort==5.13.2
lxml==5.2.1
natsort==8.4.0
onnxruntime-gpu==1.17.1
openai-whisper==20231117
prospector==1.10.3
pyannote.audio==3.1.1
pymongo==4.7.0
python-dotenv==1.0.1
torch==2.2.2
torchaudio==2.2.2
Unidecode==1.3.8
urllib3==2.2.1
